requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiofiles==23.2.1
aiohttp==3.9.1
Pillow==10.1.0
